% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_lmer.R
\name{.optim_lmer}
\alias{.optim_lmer}
\title{Optimize likelihood of lme4 for transformation parameter}
\usage{
.optim_lmer(formula, data = NULL, transform, ...)
}
\description{
Optimize likelihood of lme4 for transformation parameter
}
\keyword{internal}
