% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmm_transform_R_paket.R
\name{wlmm_transform}
\alias{wlmm_transform}
\title{Survey-weighted Linear Mixed Models under Box-Cox and Dual Transformations}
\usage{
wlmm_transform(formula, data = NULL, transform = "none",
       lambda = NULL, weights = NULL,
       iter1  = 501, iter2 = 501, MI = 500,
       tol1 = 3e-5, tol2 = 1e-5,
       trace = FALSE, nDecrease = 2, ...)
}
\arguments{
\item{formula}{Formula for the Model to be estimated. Has the shape of lmer in the package lme4.}

\item{data}{Dataframe containing the variables}

\item{transform}{Chosen transformation: "none", "box-cox" or "dual"}

\item{lambda}{Start value for transformation parameter lambda}

\item{weights}{Survey weights}

\item{iter1}{Maximum number of MCEM iterations}

\item{iter2}{Maximum number of internal optimizations within a MCEM-step}

\item{MI}{Number of importance-sampled random numbers in E-step}

\item{tol1}{Convergence criterion for MCEM algorithm}

\item{tol2}{Convergence criterion for opimizations within a MCEM-step}

\item{trace}{Print output of each MCEM-step?}

\item{nDecrease}{How often is it allowed that the estimate of LL does not increase?}

\item{...}{Additional options for model set-up}
}
\value{
A list with the following elements
\itemize{
\item model C++ output of estimation procedure:
\itemize{
\item coef Vector of fixed effects regression parameters
\item vcov_coef Fisher information of fixed effects parameters
\item lambda Transformation Parameter
\item VarCov List of random effects variance-covariance matrices
\item sigma Residual standard deviation
\item transform Chosen transformation
\item RE_mat Matrix with simulated random effects from last E-step, including importance weights
\item RE_mode Modes of the random effects
\item RE_invHess Inverse hessian of LL at mode of random effects
\item LLmod Joint maximum log-likelihood of observed data and the mode of random effects
\item LLexp Expected log-likelihood given the observed data
\item LL Estimated log-likelihood
\item convergence Has the model converged?
}
\item formula regression model
\item data Final estimate of transformation parameter
}
}
\description{
Estimates linear mixed models under Box-Cox and Dual Transformations with flexible random effects structure and survey weights.
}
\details{
For comparability between runs, set a seed due to stochastic optimization.
}
\examples{
\dontrun{
library(emdi)
data("eusilcA_smp")
mod <- wlmm_transform( eqIncome ~ gender + eqsize + cash + self_empl +
               unempl_ben + age_ben + surv_ben + sick_ben + dis_ben + rent + fam_allow +
               house_allow + cap_inv + tax_adj + (1|district), data = eusilcA_smp,
               transform = "box-cox", weights = eusilcA_smp$weight,
               MI = 500, iter1 = 250, trace = TRUE)
}
}
\references{
Burgard, Jans Pablo and Doerr, Patricia (2019). Data-driven transformations and survey-weighting for linear mixed models. Trier University. Working Paper.
}
